+++ CHANGELOG.md
@@ -89,6 +89,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [1238](https://github.com/umee-network/umee/pull/1238) Added bad debts query.
 - [1323](https://github.com/umee-network/umee/pull/1323) Oracle cli - Add validator address override option.
 - [1329](https://github.com/umee-network/umee/pull/1329) Implement MaxCollateralShare.
+- [1330](https://github.com/umee-network/umee/pull/1330) Implemented MaxSupplyUtilization.
 
 ### Improvements